Added additional check to wait for cluste operaotor api (#17)
In my initial testing, `oc get co` was already returning a few
operators. this meant that my json queries would always work. It was
only when I started testing with a cluster install which didn't even
have a working api running, I noticed that the json query vars were
failing. So added an additinal check to wait for `oc get co` to return a
valid resonpse before we start polling the cluster operators.