Adding node-selector for the openshift-cluster-ops namespace.
It is very common practice to configure a defaultNodeSelector for your
OCP cluster's Scheduler. Typically the defaultNodeSelector is set to
node-role.kubernetes.io/worker=

When the cronjob pod for the etcd-backup job spins up it will inherit
the defaultNodeSelector setting of node-role.kubernetes.io/worker=
However the etcd backup pods need to run on one of the master nodes. In
effect the cronjob pods become unschedulable.

Add a node-selector for the openshfit-cluster-ops namespace, prevent the
etcd-backup cron job from inheriting the defaultNodeSelector configured for
your cluster Scheduler.