Moving to a more deterministic way to fetch etcd image
Previously we were fetching the etcd pod in openshift-etcd and checking
the image used by the containers in the Pod. It so happens, that OCP 4.4
changed the name of the etcd pod from "etcd-member" to "etcd".

So instead going to resort to `oc adm release info` command to fetch the
image used for etcd. This should provide a more consitent way instead of
for ever trying to keep up with the name of the etcd pod.