Removing the need to DNS lookup to get master node's IP
the dig command was added to obtain the master nodes's ip from withing
the CronJob Pod. This was a work around that as put in place because the
env settings to reference the Host IP address was not working. My
original tests with env variable lookups were noot working because the
env section was not properly alliged in the yaml. The env settings are a
properly of the container. I had it one level above. It was not
erroring out when creating the pod so didn't catch it.